feat: 增添通过路径参数查看帖子详情的接口

+++ dao/mysql/community.go
@@ -23,7 +23,7 @@ func GetCommunityList() (data []*models.Community, err error) {
 }
 
 // GetCommunityByID 根据id获取社区的详情
-func GetCommunityByID(id int) (communitydetail *models.CommunityDetail, err error) {
+func GetCommunityByID(id int64) (communitydetail *models.CommunityDetail, err error) {
 	communitydetail = new(models.CommunityDetail)
 
 	sqlstr := "select community_id, community_name, introduction from community where community_id = ?"
@@ -32,11 +32,11 @@ func GetCommunityByID(id int) (communitydetail *models.CommunityDetail, err erro
 
 	if err != nil {
 		if err == sql.ErrNoRows { //特殊情况，id不存在，记录日志
-			zap.L().Warn("communit not found", zap.Int("communit_id", id))
+			zap.L().Warn("communit not found", zap.Int64("communit_id", id))
 			return nil, nil
 		}
 		// 一般情况，查询执行失败
-		zap.L().Error("query community by id failed", zap.Int("community_id", id))
+		zap.L().Error("query community by id failed", zap.Int64("community_id", id))
 
 		return nil, err
 	}
