+++ models/community.go
@@ -1,12 +1,12 @@
 package models
 
 type Community struct {
-	ID   int    `json:"id" db:"community_id"`
+	ID   int64  `json:"id" db:"community_id"`
 	Name string `json:"name" db:"community_name"`
 }
 
 type CommunityDetail struct {
-	ID           int    `json:"id" db:"community_id"`
+	ID           int64  `json:"id" db:"community_id"`
 	Name         string `json:"name" db:"community_name"`
 	Introduction string `json:"introduction" db:"introduction"`
 }
